requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
et-xmlfile==1.1.0
fake-user-agent==2.1.8
frozenlist==1.4.1
h11==0.14.0
idna==3.6
Jinja2==3.1.3
lxml==4.9.4
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.26.2
openpyxl==3.1.2
outcome==1.3.0.post0
pandas==2.1.4
pip==22.3.1
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
selenium==4.16.0
setuptools==65.5.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio-websocket==0.11.1
trio==0.23.2
tzdata==2023.3
urllib3==2.1.0
wheel==0.38.4
wsproto==1.2.0
XlsxWriter==3.1.9
yarl==1.9.4
python-dotenv==1.0.1
pyTelegramBotAPI==4.22.1
git+https://github.com/OtdelAnaliticiAlterra/TelegramBotLogger.git
